Major Changes to NPCs
NPCs will now be registered to setAlive == false if they are killed during the interaction, both by using "Patient Healthy" interaction and using the wrong treatment or in the case of using a dirty syringe. When set to isAlive() == false, the NPCs will stay gone if they died and the player is returning to the hut.